+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CMS-013] Menu admin module to manage menus and nested menu items with hierarchical tree views and link options.
 - [CMS-014] BlockType admin module to manage block configurations and localized schemas.
 - [CMS-015] Collection admin module to manage translatable system collections and metadata settings.
+- [CMS-016] Entry admin module to manage entries with multi-language translation support, custom block instances, categories/tags badges, status filtering, and scheduled publishing options.
 - [CMS-019] Language admin module to list, create, edit, delete, and set default languages.
 - [CMS-020] Setting admin module to manage translatable system configuration variables.
 - Transversal CMS sidebar section header and Lucide icons in navigation.